+++ .eslintrc.js
@@ -2,9 +2,12 @@ module.exports = {
   root: true,
   extends: [
     'airbnb-base',
-    'plugin:jest/all'
+    'plugin:jest/all',
+    "plugin:sonarjs/recommended"
   ],
-  plugins: [],
+  plugins: [
+    'sonarjs',
+  ],
   env: {
     mocha: true,
   },
@@ -27,6 +30,12 @@ module.exports = {
         maxBOF: 0,
         maxEOF: 0
       }
-    ]
+    ],
+    'sonarjs/cognitive-complexity': 1,
+    'sonarjs/no-collapsible-if': 0,
+    'sonarjs/no-extra-arguments': 0,
+    'sonarjs/no-duplicate-string': 0,
+    'sonarjs/no-identical-functions': 0,
+    'sonarjs/no-same-line-conditional': 0,
   }
 };